Succeed running test with tokio
